Move local chroma set up to utils
